softspoken + base_ot + ot_extension: drop now-unneeded casts
Track emp-tool's int64_t length-param sweep. Casts that existed only
to satisfy the previous narrower API (or to defend against int * int
overflow in pointer arithmetic that now uses int64_t natively) are
gone:

- ot_extension.h: leftover_pos_ / leftover_count_ int → int64_t to
  match the int64_t `take` / `chunk` they hold. The four (int)take /
  (int)(chunk - take) narrowing casts in drain_leftover / rcot_run
  follow.
- softspoken_ot.cpp:
  - `io->send_block(t_buf.data(), static_cast<int>(n) * 2)` → drop the
    cast; send_block takes int64_t.
  - Same on the receiver-side recv_block call.
  - `hash.put(s_buf.data(), Q * 2 * (int)sizeof(block))` → drop the
    (int) narrowing; Hash::put takes int64_t.
  - Pointer arithmetic `static_cast<size_t>(i) * k * bs` → plain
    `i * k * bs`. bs is int64_t, so int * int * int64_t auto-widens.
  - Same shape on `d_i = d_bufs + (i - 1) * bs`.
- sfvole_butterfly.h: every `(size_t)` cast on pointer-arith
  expressions (`b * v_stride`, `t * L::N`, `bs + t0`, `n*(size_t)bs`)
  is now redundant — bs / v_stride are int64_t, so widening happens
  through the multiplication.
- base_ot/csw.h: drop the `(int)` cast on `hash_for_block`'s size
  argument in H_pad and H_aggregate (hash_for_block takes int64_t).

trace_equiv passes (wire bytes bit-identical). All 8 emp-ot tests
pass; bench_softspoken_rcot all 12 configs print "Tests passed."

+++ emp-ot/base_ot/csw.h
@@ -84,7 +84,7 @@ class OTCSW : public OT { public:
 		memcpy(buf + 1 + sizeof(block), &i, sizeof(int64_t));
 		P.to_bin(buf + 1 + sizeof(block) + sizeof(int64_t), plen);
 		return Hash::hash_for_block(buf,
-			(int)(1 + sizeof(block) + sizeof(int64_t) + plen));
+			1 + sizeof(block) + sizeof(int64_t) + plen);
 	}
 
 	// H_3(sid, x) → block. x is a single block (used both for hashing
@@ -106,7 +106,7 @@ class OTCSW : public OT { public:
 		buf[0] = '4';
 		memcpy(buf.data() + 1, &sid, sizeof(block));
 		memcpy(buf.data() + 1 + sizeof(block), hs, (size_t)ell * sizeof(block));
-		return Hash::hash_for_block(buf.data(), (int)hlen);
+		return Hash::hash_for_block(buf.data(), hlen);
 	}
 
 	// ----- Sender side. Plays the OT sender role (S in the paper). -----
